chore: rewrite osm base loaders
